+++ NEWS.md
@@ -1,6 +1,22 @@
-# RWDataPlyr 0.6.2.9000
+# RWDataPlyr 0.6.3
 
-*In development*
+*Releases February 28, 2020*
+
+## Bug Fixes
+
+* Added additional test for rdf files that are not found when using `rw_scen_aggregate()` (#97)
+* Fixed the variable name assigned to Mead.Pool Elevation inside example `rwd_agg` that is created by `rwd_agg_template(examples = TRUE)`(#94)
+* When aggregating slots, if the specified period is "eocy"", then the summary must be `NA` (#101)
+* Now check directory exists before checking individual files exist in `rdf_aggregate()` (#108)
+
+## Documentation
+
+* Updated documentation for the `rwslot_*()` functions
+
+## Under the Hood Changes
+
+* Changed test that is expecting a data.frame with factors so that `rwd_agg()` fails. Explicitly stating that strings should be factors to work with R v4.0.0.
+* Switched all `rwd_agg(read.csv())` calls to `read_rwd_agg()`
 
 # RWDataPlyr 0.6.2
 
@@ -10,7 +26,7 @@
 
 * The `oFile` argument in `getDataForAllScens()` now defaults to `NULL`, so that this function conforms to CRAN policies regarding not writing to the user's home file space by default. This should not cause any backwards compatability issues since all older code will explicitly specify the `oFile` argument.
 * The default `file` argument in `rwd_agg_template()` is now empty, so the user must specify the file explicitly for it to be created, also to conform to the same CRAN policy.
-* Tests, examples, and the vignetter were updated to respect this policy, by only writing to the `tempdir()`, when necessary.
+* Tests, examples, and the vignette were updated to respect this policy, by only writing to the `tempdir()`, when necessary.
 
 # RWDataPlyr 0.6.1